refactor: *StockCorrectionBody
